app-backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.35
pydantic[email]==2.9.0
passlib[bcrypt]==1.7.4
PyJWT==2.9.0
pandas==2.2.3
numpy==1.26.4
scikit-learn==1.5.2
joblib==1.4.2
lightgbm==4.5.0
vaderSentiment==3.3.2
bcrypt==4.2.0
